feat(uglifyjs): renamed flow to uglify for consistency
fix #409
