+++ lib/config/uglify.js
