+++ lib/configwriter.js
@@ -23,7 +23,7 @@ var deepMerge = function (origCfg, cfg) {
 //
 // flow is an array of strings that contains name of each steps, in order.
 // Each of the steps is corresponding to a processor that will be used on the files
-// enclosed in blocks (for example, concat, uglifyjs).
+// enclosed in blocks (for example, concat, uglify).
 // Note that a corresponding file implementing a dedicated writer needs to be present
 // under the same name, in config directory.
 //
@@ -32,11 +32,11 @@ var deepMerge = function (origCfg, cfg) {
 //
 // For example:
 //
-//   var w = new ConfigWriter( ['concat', 'uglifyjs'], [], {input: 'app', staging: '.tmp', dest: 'dist'});
+//   var w = new ConfigWriter( ['concat', 'uglify'], [], {input: 'app', staging: '.tmp', dest: 'dist'});
 //
 //   We'll:
-//     - have files config/concat.js and config/uglifyjs.js implementing each config writer
-//     - Use .tmp for the output of concat and input of uglifyjs
+//     - have files config/concat.js and config/uglify.js implementing each config writer
+//     - Use .tmp for the output of concat and input of uglify
 //     - Deliver for each block the requested file under dist directory
 //
 //
@@ -52,8 +52,8 @@ var ConfigWriter = module.exports = function (flow, dirs) {
   this.destinations = {};
 
   // We need to create all the needed config writers, given them their output directory
-  // E.g, if we do have the flow concat | uglifyjs, the output dir will be .tmp/concat and dist
-  // if the flow is concat | uglifyjs | foo, the dirs will be .tmp/concat, .tmp/uglifyjs, dist
+  // E.g, if we do have the flow concat | uglify, the output dir will be .tmp/concat and dist
+  // if the flow is concat | uglify | foo, the dirs will be .tmp/concat, .tmp/uglify, dist
   flow.blockTypes().forEach(function (blockType) {
     self.steps[blockType] = [];
     self.postprocessors[blockType] = [];
