+++ test/test-config-uglifyjs.js
@@ -1,6 +1,6 @@
 'use strict';
 var assert = require('assert');
-var uglifyjsConfig = require('../lib/config/uglifyjs.js');
+var uglifyConfig = require('../lib/config/uglify.js');
 var path = require('path');
 
 var block = {
@@ -20,22 +20,22 @@ var block = {
   ]
 };
 
-describe('Uglifyjs config write', function () {
+describe('Uglify config write', function () {
   it('should have a correct name', function () {
-    assert.equal(uglifyjsConfig.name, 'uglify');
+    assert.equal(uglifyConfig.name, 'uglify');
   });
 
   it('should use the input files correctly', function () {
     var ctx = {
       inDir: 'zzz',
       inFiles: ['foo.js', 'bar.js', 'baz.js'],
-      outDir: 'tmp/uglifyjs',
+      outDir: 'tmp/uglify',
       outFiles: []
     };
-    var cfg = uglifyjsConfig.createConfig(ctx, block);
+    var cfg = uglifyConfig.createConfig(ctx, block);
     assert.ok(cfg.files);
     assert.equal(cfg.files.length, 3);
-    var dests = ['tmp/uglifyjs/foo.js', 'tmp/uglifyjs/bar.js', 'tmp/uglifyjs/baz.js'];
+    var dests = ['tmp/uglify/foo.js', 'tmp/uglify/bar.js', 'tmp/uglify/baz.js'];
     var srcs = ['zzz/foo.js', 'zzz/bar.js', 'zzz/baz.js'];
 
     cfg.files.forEach(function (files, idx) {
@@ -56,7 +56,7 @@ describe('Uglifyjs config write', function () {
       outFiles: [],
       last: true
     };
-    var cfg = uglifyjsConfig.createConfig(ctx, block);
+    var cfg = uglifyConfig.createConfig(ctx, block);
     assert.ok(cfg.files);
     assert.equal(cfg.files.length, 1);
     var files = cfg.files[0];
@@ -68,12 +68,12 @@ describe('Uglifyjs config write', function () {
 
   // it('should allow for concatenation', function () {
   //   var requirejsConfig = {};
-  //   var co = new uglifyjsConfig( 'tmp/uglifyjs', requirejsConfig );
+  //   var co = new uglifyConfig( 'tmp/uglify', requirejsConfig );
   //   var ctx = { inFiles: ['foo.js', 'bar.js', 'baz.js'], outFiles: []};
   //   var cfg = co.createConfig(ctx, block);
-  //   assert.ok(cfg['tmp/uglifyjs/foo.js']);
-  //   assert.deepEqual(cfg['tmp/uglifyjs/scripts/site.js'], ['foo.js', 'bar.js', 'baz.js']);
-  //   assert.deepEqual(ctx.outFiles, ['tmp/uglifyjs/scripts/site.js']);
+  //   assert.ok(cfg['tmp/uglify/foo.js']);
+  //   assert.deepEqual(cfg['tmp/uglify/scripts/site.js'], ['foo.js', 'bar.js', 'baz.js']);
+  //   assert.deepEqual(ctx.outFiles, ['tmp/uglify/scripts/site.js']);
   // });
 
 });
